Add run gc after trial setting
