+++ Tunny/Solver/EndState.cs
@@ -1,0 +1,12 @@
+namespace Tunny.Solver
+{
+    public enum EndState
+    {
+        AllTrialCompleted,
+        Timeout,
+        StoppedByUser,
+        DirectionNumNotMatch,
+        UseExitStudyWithoutLoading,
+        Error
+    }
+}
